Moves helper methods into separate file for backward compatability
